+++ api/src/main/java/org/mandulis/mts/auth/AttachmentController.java
@@ -1,15 +1,12 @@
 package org.mandulis.mts.controller;
 
 import org.mandulis.mts.dto.AttachmentDto;
-import org.mandulis.mts.exception.ProcessingMultipartFileException;
+import org.mandulis.mts.exception.EmptyAttachmentException;
 import org.mandulis.mts.service.AttachmentService;
 import org.mandulis.mts.service.AttachmentValidationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.io.InputStream;
 
 @RestController
 @RequestMapping("/api/v1/user/tickets")
@@ -28,18 +25,18 @@ class AttachmentController {
             @PathVariable("ticketId") Long ticketId,
             @RequestParam(value = "file") MultipartFile multipartFile
     ) {
-        var dto = new AttachmentDto(multipartFile.getOriginalFilename(), ticketId, toInputStream(multipartFile));
-        attachmentValidationService.validateAttachment(multipartFile);
-        attachmentService.save(dto);
+        var attachmentDto = new AttachmentDto(multipartFile, ticketId);
+        validateMultipart(multipartFile);
+        attachmentValidationService.validateAttachment(attachmentDto);
+        attachmentService.save(attachmentDto);
         return ResponseEntity.status(201).build();
     }
 
-    private InputStream toInputStream(MultipartFile multipartFile) {
-        try {
-            return multipartFile.getInputStream();
-        } catch (IOException e) {
-            throw new ProcessingMultipartFileException("File invalid.", e);
+    private void validateMultipart(MultipartFile multipartFile) {
+        if(multipartFile.isEmpty()) {
+            throw new EmptyAttachmentException();
         }
     }
+
 
 }
