+++ api/src/main/java/org/mandulis/mts/service/AttachmentValidationService.java
@@ -1,25 +1,15 @@
 package org.mandulis.mts.service;
 
-import lombok.RequiredArgsConstructor;
 import org.mandulis.mts.dto.AttachmentDto;
 import org.mandulis.mts.exception.AttachmentWithoutNameException;
-import org.mandulis.mts.exception.EmptyAttachmentException;
-import org.mandulis.mts.exception.TicketNotFoundException;
-import org.mandulis.mts.repository.TicketRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
-@RequiredArgsConstructor
 public class AttachmentValidationService {
 
-    public void validateAttachment(MultipartFile multipartFile) {
-        if(multipartFile.getOriginalFilename() == null || multipartFile.getOriginalFilename().isEmpty()) {
+    public void validateAttachment(AttachmentDto attachmentDto) {
+        if (attachmentDto.fileName == null || attachmentDto.fileName.isEmpty()) {
             throw new AttachmentWithoutNameException();
-        }
-
-        if(multipartFile.isEmpty()) {
-            throw new EmptyAttachmentException();
         }
     }
 }
