--- api/src/main/java/org/mandulis/mts/service/StorageService.java
@@ -1,9 +1,0 @@
-package org.mandulis.mts.service;
-
-import java.io.InputStream;
-import java.nio.file.Path;
-
-public interface StorageService {
-
-    Path store(String filename, InputStream file);
-}
