+++ api/src/main/resources/application.properties
@@ -23,3 +23,8 @@ jwt.secret=${JWT_SECRET}
 # upload limits
 spring.servlet.multipart.max-file-size=128KB
 spring.servlet.multipart.max-request-size=128KB
+
+minio.endpoint=${MINIO.ENDPOINT}
+minio.username=${MINIO.USERNAME}
+minio.password=${MINIO.PASSWORD}
+minio.bucketName=${MINIO.BUCKET_NAME}
